Generator, Rarity, Saves, Shield, Level etc.
Rooms more efficiently connected.
Rarity corresponds to stat count, there are now colors for rarity.
Many players can now save their progress.
Working on a shield game mechanic.
Levels can now have some level-bounded loot, treasure rooms WIP.

+++ Assets/Level/Control/Item System/StatPattern.cs
@@ -21,7 +21,7 @@ public class StatPattern : ScriptableObject
     /// <summary>
     /// o kolik se zvýší hodnota statu za každou úroveň předmětu nad 1
     /// </summary>
-    public int incrementPerLvl;
+    public float incrementPerLvl;
     /// <summary>
     /// možné typy itemů, na kterých se stat může vygenerovat
     /// </summary>
